+++ test/expect/test__cmd_name_with_underscore.ml
@@ -1,0 +1,7 @@
+module Command = Cmdlang.Command
+
+let return =
+  let open Command.Std in
+  let+ () = Arg.return () in
+  ()
+;;